Refactoring of the SFTP/WebDAV component (#1)
This commit contains quite a big change in terms of refactoring. The old problematic code was dropped and replaced in favor of ES6/ES7 one which handles the whole process in a better way. A big focus was made on memory consumption. And the files in the SFTP processing are handled via streams and the promises are handled in the series. The logic for the WebDAV stayed the same, just the things are written more nicely and in more consistent way. There are also better logs and documentation was little updated.